# PERF: releasing the GIL, #8882