… pandas/hashtable.pyx
… pandas/src/generate_code.py
… pandas/src/generated.pyx
… vb_suite/suite.py