closes #8882
 
This is only implemened ATM for:

```
grp = df.groupby('key1')['data']
grp.sum()
```

I am isolating this because in this example computing the group indexers is actually quite a bit more expensive that actually doing the sum itself.


consecutive executions show a linear scaling as one would expect
```
In [20]: %timeit g2()
100 loops, best of 3: 7.72 ms per loop

In [21]: %timeit g4()
100 loops, best of 3: 14.4 ms per loop

In [22]: %timeit g8()
10 loops, best of 3: 29.2 ms per loop
```

Same but executing the same function in the indicated number of threads
(the same code run in master with the ``test_parallel`` decorator is very close to the above)
```
In [23]: %timeit pg2()
100 loops, best of 3: 4.95 ms per loop

In [24]: %timeit pg4()
100 loops, best of 3: 6.48 ms per loop

In [25]: %timeit pg8()
100 loops, best of 3: 10.9 ms per loop
```

Pls give this a test in some downstream systems, e.g. ``dask`` would be a good test
cc @mrocklin
@cpcloud