@mrocklin 2015-05-22
What was the deal with `NA` and `nan`.  These were Python objects?