@jreback 2015-05-22
apparently doing something

```
cdef ndarray[double_t] foo

foo = np.empty(10,dtype='float64')

foo[0] = np.nan
```

actually assigns a python object (and does inference and such), so that you MUST have the gil,
but casting to a double seems to work though (as its now c-level)