@mrocklin 2015-05-29
Cool.  I'll try to run a few different groupby operations with dask.dataframe sometime early next week and report back.  I'm curious what were the operations that needed to be released?  Was there a core set of functionality or was it a large volume of trivial work?