@jreback 2015-05-29
TL;DR.

So lots of 'boilerplate' changes in ``generated.pyx``, which is basically all of the algorithms for groupby ops themselves (e.g. sum,mean). These were straightforward insertions of ``with nogil``, just needed to avoid typing the ``object`` ones (which for the most part was easy; though a bit tricky on some non-groupby routines, e.g. see #10213).

However, the big speeds occur once I fixed the ``hashtable.pyx`` factorizer (e.g. the ``get_labels`` routine). The trivial soln actually produces a very odd perf issue (see [here](https://groups.google.com/forum/#!topic/cython-users/jtZ0FOGX2vU)), which is a bug / documentation issue.

In effect, you are using a ``nogil`` function that happens to have a ``with gil`` block WITHIN it, then the function is way slower than if you inline the exact same code.

I 'fixed' this issue by adding some additional c-structures to hold data (basically I changed the low-level implementation); these can be easily passed around; downside is that are a little 'messy' in the code department.

note to @shoyer I didn't actually bypass the realloc of memory (e.g. ``.resize``) of numpy arrays. though I do have a soln to that (basically manage a memory block myself, then turning that into a numpy array at the very end (in ``to_array``), easy enough, but was running into some compile issues, so dropped it in the end).