@mrocklin 2015-06-17
I think that this is awesome and am afraid of it going stale.  Is there something blocking it from being merged into master?