@jreback 2015-06-17
actually I do need to add a doc-note to the whatsnew (and maybe a small mention in the enhancingperf section as well).