@mrocklin 2015-06-17
Hrm, yes, unfortunate that you can't manipulate the lock directly