@jreback 2015-06-17
yep, would be nice to put in something along the lines of an ``#IFDEF``, but in reality the block is actually doing 'work', so maybe not easy/possible.