@mrocklin 2015-06-26
Thought I'd caress this PR with a gentle ping.  

*ping*