@mrocklin 2015-06-26
Kind and loving harassment, yes.