@jreback 2015-06-26
any commentary @jorisvandenbossche @shoyer @cpcloud @TomAugspurger 

going to have a followup for more GIL stuff. Its actually tricky making sure its doing (and you are measure exactly the right stuff)