@mrocklin 2015-06-30
Woohoo!

*Pulls from master, checks benchmarks.*