@mrocklin 2015-06-30
Best I have to offer at the moment are the slowly growing [`dask.dataframe` docs](http://dask.pydata.org/en/latest/dataframe.html)